Fix #176: Prepare semver 3.0.0-alpha0
* Remove targets py27, py34, py35, and pypy from `tox.ini`
* Update `README.rst` and remove anything related to Python2
  Mention maintenance branch `maint/v2`
* `setup.py`
  - Update Trove classifiers
  - Require now Python >=3.6.*
  - Remove Tox and Clean classes, try to make it
    as simple as possible
  - Extract metadata directly from source (affects all the __version__,
    __author__ etc. variables)
* `setup.cfg`
  - Add pycodestyle section
* `semver.py`
  - Change version number to "3.0.0-alpha0"
  - Remove old code related to Python2
  - Adjust Python2 vs. Python3 str/bytes
  - Add wheel as another test requirement
* `tox.ini`
  - Remove py27, py34, and py35 (out of maintenance)
  - Add docs to default testenv
  - Remove --universal from bdist_wheel
* Split test suite into separate files under tests/ dir
* Travis: Remove old versions, integrate new
  - Remove 2.7, 3.4, 3.5, and pypy
  - Integrate 3.8, 3.9-dev and nightly builds
  - Allow nightly to fail
* Add .editorconfig to have a consistent editor setup
