+++ CHANGELOG.rst
@@ -39,7 +39,6 @@ Features
 
 n/a
 
-
 Bug Fixes
 ---------
