+++ docs/_static/css/semver.css
@@ -1,0 +1,20 @@
+/*
+*/
+
+.py.method {
+    padding-top: 0.25em;
+    padding-bottom: 1.25em;
+    border-top: 1px solid #EEE;
+}
+
+.py.function{
+    padding-top: 1.25em;
+}
+
+.related.bottom {
+    margin-top: 1em;
+}
+
+nav#rellinks {
+    float: left;
+}
