+++ docs/_templates/layout.html
@@ -1,0 +1,4 @@
+{#
+  Import the theme's layout.
+#}
+{% extends "!layout.html" %}
