+++ setup.cfg
@@ -1,6 +1,6 @@
 [tool:pytest]
 norecursedirs = .git build .env/ env/ .pyenv/ .tmp/ .eggs/
-testpaths = . docs
+testpaths = tests docs
 filterwarnings =
     ignore:Function 'semver.*:DeprecationWarning
 addopts =
@@ -24,3 +24,15 @@ exclude =
     dist
     docs
     conftest.py
+
+[pycodestyle]
+count = False
+# ignore = E226,E302,E41
+max-line-length = 88
+statistics = True
+exclude =
+    .env,
+    .eggs,
+    .tox,
+    .git,
+    docs
