+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
 envlist =
     flake8
-    py{27,34,35,36,37}
-    pypy
+    py{36,37,38,39,310}
+    docs
 
 [testenv]
-description = Run test suite
+description = Run test suite for {basepython}
 whitelist_externals = make
 commands = pytest {posargs:}
 deps =
@@ -58,7 +58,6 @@ deps = sphinx
 skip_install = true
 commands = make -C docs man
 
-
 [testenv:prepare-dist]
 description = Prepare for TestPyPI
 basepython = python3
@@ -66,5 +65,5 @@ deps =
     wheel
     twine
 commands =
-    python3 setup.py sdist bdist_wheel --universal
+    python3 setup.py sdist bdist_wheel
     twine check dist/*
